+++ src/main/java/org/gitlab4j/api/BoardsApi.java
@@ -12,10 +12,10 @@ import org.gitlab4j.api.models.BoardList;
 
 /**
  * This class implements the client side API for the GitLab Issue Boards API calls.
- * 
- * NOTE: If a user is not a member of a group and the group is private, 
+ *
+ * NOTE: If a user is not a member of a group and the group is private,
  *       a GET request on that group will result to a 404 status code.
- * 
+ *
  * @see <a href="https://docs.gitlab.com/ce/api/boards.html">GitLab Issue Boards API Documentaion</a>
  */
 public class BoardsApi extends AbstractApi {
@@ -29,7 +29,7 @@ public class BoardsApi extends AbstractApi {
      *
      * <pre><code>GitLab Endpoint: GET /projects/:id/boards</code></pre>
      *
-     * @param projectIdOrPath the project in the form of an Integer(ID), String(path), or Project instance
+     * @param projectIdOrPath the project in the form of an Long(ID), String(path), or Project instance
      * @return a list of project's issue boards
      * @throws GitLabApiException if any exception occurs
      */
@@ -42,7 +42,7 @@ public class BoardsApi extends AbstractApi {
      *
      * <pre><code>GitLab Endpoint: GET /projects/:id/boards</code></pre>
      *
-     * @param projectIdOrPath the project in the form of an Integer(ID), String(path), or Project instance
+     * @param projectIdOrPath the project in the form of an Long(ID), String(path), or Project instance
      * @param page the page to get
      * @param perPage the number of items per page
      * @return a list of project's Boards in the specified range
@@ -59,7 +59,7 @@ public class BoardsApi extends AbstractApi {
      *
      * <pre><code>GitLab Endpoint: GET /projects/:id/boards</code></pre>
      *
-     * @param projectIdOrPath the project in the form of an Integer(ID), String(path), or Project instance
+     * @param projectIdOrPath the project in the form of an Long(ID), String(path), or Project instance
      * @param itemsPerPage the number of items per page
      * @return a Pager of project's issue boards
      * @throws GitLabApiException if any exception occurs
@@ -74,7 +74,7 @@ public class BoardsApi extends AbstractApi {
      *
      * <pre><code>GitLab Endpoint: GET /projects/:id/boards</code></pre>
      *
-     * @param projectIdOrPath the project in the form of an Integer(ID), String(path), or Project instance
+     * @param projectIdOrPath the project in the form of an Long(ID), String(path), or Project instance
      * @return a Stream of project's issue boards
      * @throws GitLabApiException if any exception occurs
      */
@@ -87,13 +87,13 @@ public class BoardsApi extends AbstractApi {
      *
      * <pre><code>GitLab Endpoint: GET /projects/:id/boards/:board_id</code></pre>
      *
-     * @param projectIdOrPath the project in the form of an Integer(ID), String(path), or Project instance
+     * @param projectIdOrPath the project in the form of an Long(ID), String(path), or Project instance
      * @param boardId the ID of the board
      * @return a Board instance for the specified board ID
      * @throws GitLabApiException if any exception occurs
      */
-    public Board getBoard(Object projectIdOrPath, Integer boardId) throws GitLabApiException {
-        Response response = get(Response.Status.OK, null, 
+    public Board getBoard(Object projectIdOrPath, Long boardId) throws GitLabApiException {
+        Response response = get(Response.Status.OK, null,
                 "projects", getProjectIdOrPath(projectIdOrPath), "boards", boardId);
         return (response.readEntity(Board.class));
     }
@@ -101,13 +101,13 @@ public class BoardsApi extends AbstractApi {
     /**
      * Get an issue board as an Optional instance.
      *
-     * <pre><code>GitLab Endpoint: GET /projects/:id/boards/:board_id</code></pre> 
+     * <pre><code>GitLab Endpoint: GET /projects/:id/boards/:board_id</code></pre>
      *
-     * @param projectIdOrPath the project in the form of an Integer(ID), String(path), or Project instance
+     * @param projectIdOrPath the project in the form of an Long(ID), String(path), or Project instance
      * @param boardId the ID of the board
      * @return the Board instance for the specified board ID as an Optional instance
      */
-    public Optional<Board> getOptionalBoard(Object projectIdOrPath, Integer boardId) {
+    public Optional<Board> getOptionalBoard(Object projectIdOrPath, Long boardId) {
         try {
             return (Optional.ofNullable(getBoard(projectIdOrPath, boardId)));
         } catch (GitLabApiException glae) {
@@ -122,7 +122,7 @@ public class BoardsApi extends AbstractApi {
      *
      * <pre><code>GitLab Endpoint: POST /projects/:id/boards</code></pre>
      *
-     * @param projectIdOrPath the project in the form of an Integer(ID), String(path), or Project instance
+     * @param projectIdOrPath the project in the form of an Long(ID), String(path), or Project instance
      * @param name the name for the new board
      * @return the created Board instance
      * @throws GitLabApiException if any exception occurs
@@ -140,7 +140,7 @@ public class BoardsApi extends AbstractApi {
      *
      * <pre><code>GitLab Endpoint: PUT /projects/:id/boards/:board_id</code></pre>
      *
-     * @param projectIdOrPath the project in the form of an Integer(ID), String(path), or Project instance, required
+     * @param projectIdOrPath the project in the form of an Long(ID), String(path), or Project instance, required
      * @param boardId the ID of the board, required
      * @param name the new name of the board, optional (can be null)
      * @param assigneeId the assignee the board should be scoped to, optional (can be null)
@@ -150,8 +150,8 @@ public class BoardsApi extends AbstractApi {
      * @return the updated Board instance
      * @throws GitLabApiException if any exception occurs
      */
-    public BoardList updateBoard(Object projectIdOrPath, Integer boardId, String name, 
-            Integer assigneeId, Integer milestoneId, String labels, Integer weight) throws GitLabApiException {
+    public BoardList updateBoard(Object projectIdOrPath, Long boardId, String name,
+            Long assigneeId, Long milestoneId, String labels, Integer weight) throws GitLabApiException {
         GitLabApiForm formData = new GitLabApiForm()
                 .withParam("name", name)
                 .withParam("assignee_id", assigneeId)
@@ -170,11 +170,11 @@ public class BoardsApi extends AbstractApi {
      *
      * <pre><code>GitLab Endpoint: DELETE /projects/:id/boards/:board_id</code></pre>
      *
-     * @param projectIdOrPath the project in the form of an Integer(ID), String(path), or Project instance
+     * @param projectIdOrPath the project in the form of an Long(ID), String(path), or Project instance
      * @param boardId the ID of the board
      * @throws GitLabApiException if any exception occurs
      */
-    public void deleteBoard(Object projectIdOrPath, Integer boardId) throws GitLabApiException {
+    public void deleteBoard(Object projectIdOrPath, Long boardId) throws GitLabApiException {
         delete(Response.Status.NO_CONTENT, null, "projects", getProjectIdOrPath(projectIdOrPath), "boards", boardId);
     }
 
@@ -183,12 +183,12 @@ public class BoardsApi extends AbstractApi {
      *
      * <pre><code>GitLab Endpoint: GET /projects/:id/boards/:board_id/lists</code></pre>
      *
-     * @param projectIdOrPath the project in the form of an Integer(ID), String(path), or Project instance
+     * @param projectIdOrPath the project in the form of an Long(ID), String(path), or Project instance
      * @param boardId the ID of the board
      * @return a list of the issue board's lists
      * @throws GitLabApiException if any exception occurs
      */
-    public List<BoardList> getBoardLists(Object projectIdOrPath, Integer boardId) throws GitLabApiException {
+    public List<BoardList> getBoardLists(Object projectIdOrPath, Long boardId) throws GitLabApiException {
         return (getBoardLists(projectIdOrPath, boardId, getDefaultPerPage()).all());
     }
 
@@ -198,14 +198,14 @@ public class BoardsApi extends AbstractApi {
      *
      * <pre><code>GitLab Endpoint: GET /projects/:id/boards/:board_id/lists</code></pre>
      *
-     * @param projectIdOrPath the project in the form of an Integer(ID), String(path), or Project instance
+     * @param projectIdOrPath the project in the form of an Long(ID), String(path), or Project instance
      * @param boardId the ID of the board
      * @param page the page to get
      * @param perPage the number of Boards per page
      * @return a list of the issue board's lists in the specified range
      * @throws GitLabApiException if any exception occurs
      */
-    public List<BoardList> getBoardLists(Object projectIdOrPath, Integer boardId, int page, int perPage) throws GitLabApiException {
+    public List<BoardList> getBoardLists(Object projectIdOrPath, Long boardId, int page, int perPage) throws GitLabApiException {
         Response response = get(javax.ws.rs.core.Response.Status.OK, getPageQueryParams(page, perPage),
                 "projects", getProjectIdOrPath(projectIdOrPath), "boards", boardId, "lists");
         return (response.readEntity(new GenericType<List<BoardList>>() {}));
@@ -216,13 +216,13 @@ public class BoardsApi extends AbstractApi {
      *
      * <pre><code>GitLab Endpoint: GET /projects/:id/boards/:board_id/lists</code></pre>
      *
-     * @param projectIdOrPath the project in the form of an Integer(ID), String(path), or Project instance
+     * @param projectIdOrPath the project in the form of an Long(ID), String(path), or Project instance
      * @param boardId the ID of the board
      * @param itemsPerPage the number of Board instances that will be fetched per page
      * @return a Pager of the issue board's lists
      * @throws GitLabApiException if any exception occurs
      */
-    public Pager<BoardList> getBoardLists(Object projectIdOrPath, Integer boardId, int itemsPerPage) throws GitLabApiException {
+    public Pager<BoardList> getBoardLists(Object projectIdOrPath, Long boardId, int itemsPerPage) throws GitLabApiException {
         return (new Pager<BoardList>(this, BoardList.class, itemsPerPage, null,
                 "projects", getProjectIdOrPath(projectIdOrPath), "boards", boardId, "lists"));
     }
@@ -232,12 +232,12 @@ public class BoardsApi extends AbstractApi {
      *
      * <pre><code>GitLab Endpoint: GET /projects/:id/boards/:board_id/lists</code></pre>
      *
-     * @param projectIdOrPath the project in the form of an Integer(ID), String(path), or Project instance
+     * @param projectIdOrPath the project in the form of an Long(ID), String(path), or Project instance
      * @param boardId the ID of the board
      * @return a Stream of the issue board's lists
      * @throws GitLabApiException if any exception occurs
      */
-    public Stream<BoardList> getBoardsListsStream(Object projectIdOrPath, Integer boardId) throws GitLabApiException {
+    public Stream<BoardList> getBoardsListsStream(Object projectIdOrPath, Long boardId) throws GitLabApiException {
         return (getBoardLists(projectIdOrPath, boardId, getDefaultPerPage()).stream());
     }
 
@@ -246,14 +246,14 @@ public class BoardsApi extends AbstractApi {
      *
      * <pre><code>GitLab Endpoint: GET /projects/:id/boards/:board_id/lists/:list_id</code></pre>
      *
-     * @param projectIdOrPath the project in the form of an Integer(ID), String(path), or Project instance
+     * @param projectIdOrPath the project in the form of an Long(ID), String(path), or Project instance
      * @param boardId the ID of the board
      * @param listId the ID of the board lists to get
      * @return a BoardList instance for the specified board ID and list ID
      * @throws GitLabApiException if any exception occurs
      */
-    public BoardList getBoardList(Object projectIdOrPath, Integer boardId, Integer listId) throws GitLabApiException {
-        Response response = get(Response.Status.OK, null, 
+    public BoardList getBoardList(Object projectIdOrPath, Long boardId, Long listId) throws GitLabApiException {
+        Response response = get(Response.Status.OK, null,
                 "projects", getProjectIdOrPath(projectIdOrPath), "boards", boardId, "lists", listId);
         return (response.readEntity(BoardList.class));
     }
@@ -261,14 +261,14 @@ public class BoardsApi extends AbstractApi {
     /**
      * Get a single issue board list as an Optional instance.
      *
-     * <pre><code>GitLab Endpoint: GET /projects/:id/boards/:board_id/lists/:list_id</code></pre> 
+     * <pre><code>GitLab Endpoint: GET /projects/:id/boards/:board_id/lists/:list_id</code></pre>
      *
-     * @param projectIdOrPath the project in the form of an Integer(ID), String(path), or Project instance
+     * @param projectIdOrPath the project in the form of an Long(ID), String(path), or Project instance
      * @param boardId the ID of the board
      * @param listId the ID of the board lists to get
      * @return a BoardList instance for the specified board ID and list ID as an Optional instance
      */
-    public Optional<BoardList> getOptionalBoardList(Object projectIdOrPath, Integer boardId, Integer listId) {
+    public Optional<BoardList> getOptionalBoardList(Object projectIdOrPath, Long boardId, Long listId) {
         try {
             return (Optional.ofNullable(getBoardList(projectIdOrPath, boardId, listId)));
         } catch (GitLabApiException glae) {
@@ -279,15 +279,15 @@ public class BoardsApi extends AbstractApi {
     /**
      * Creates a new Issue Board list.
      *
-     * <pre><code>GitLab Endpoint: POST /projects/:id/boards/:board_id/lists</code></pre> 
+     * <pre><code>GitLab Endpoint: POST /projects/:id/boards/:board_id/lists</code></pre>
      *
-     * @param projectIdOrPath the project in the form of an Integer(ID), String(path), or Project instance
+     * @param projectIdOrPath the project in the form of an Long(ID), String(path), or Project instance
      * @param boardId the ID of the board
      * @param labelId the ID of the label
      * @return the created BoardList instance
      * @throws GitLabApiException if any exception occurs
      */
-    public BoardList createBoardList(Object projectIdOrPath, Integer boardId, Integer labelId) throws GitLabApiException {
+    public BoardList createBoardList(Object projectIdOrPath, Long boardId, Long labelId) throws GitLabApiException {
         GitLabApiForm formData = new GitLabApiForm()
                 .withParam("label_id", labelId, true);
         Response response = post(Response.Status.CREATED, formData, "projects", getProjectIdOrPath(projectIdOrPath), "boards", boardId, "lists");
@@ -297,16 +297,16 @@ public class BoardsApi extends AbstractApi {
     /**
      * Updates an existing Issue Board list. This call is used to change list position.
      *
-     * <pre><code>GitLab Endpoint: PUT /projects/:id/boards/:board_id/lists/:list_id</code></pre> 
+     * <pre><code>GitLab Endpoint: PUT /projects/:id/boards/:board_id/lists/:list_id</code></pre>
      *
-     * @param projectIdOrPath the project in the form of an Integer(ID), String(path), or Project instance
+     * @param projectIdOrPath the project in the form of an Long(ID), String(path), or Project instance
      * @param boardId the ID of the board
      * @param listId the ID of the list
      * @param position the new position for the list
      * @return the updated BoardList instance
      * @throws GitLabApiException if any exception occurs
      */
-    public BoardList updateBoardList(Object projectIdOrPath, Integer boardId, Integer listId, Integer position) throws GitLabApiException {
+    public BoardList updateBoardList(Object projectIdOrPath, Long boardId, Long listId, Integer position) throws GitLabApiException {
         GitLabApiForm formData = new GitLabApiForm().withParam("position", position, true);
         Response response = putWithFormData(Response.Status.OK, formData,
                 "projects", getProjectIdOrPath(projectIdOrPath), "boards", boardId, "lists", listId);
@@ -316,14 +316,14 @@ public class BoardsApi extends AbstractApi {
     /**
      * Soft deletes an existing Issue Board list. Only for admins and project owners.
      *
-     * <pre><code>GitLab Endpoint: DELETE /projects/:id/boards/:board_id/lists/:list_id</code></pre> 
+     * <pre><code>GitLab Endpoint: DELETE /projects/:id/boards/:board_id/lists/:list_id</code></pre>
      *
-     * @param projectIdOrPath the project in the form of an Integer(ID), String(path), or Project instance
+     * @param projectIdOrPath the project in the form of an Long(ID), String(path), or Project instance
      * @param boardId the ID of the board
      * @param listId the ID of the list
      * @throws GitLabApiException if any exception occurs
      */
-    public void deleteBoardList(Object projectIdOrPath, Integer boardId, Integer listId) throws GitLabApiException {
+    public void deleteBoardList(Object projectIdOrPath, Long boardId, Long listId) throws GitLabApiException {
         delete(Response.Status.NO_CONTENT, null, "projects", getProjectIdOrPath(projectIdOrPath), "boards", boardId, "lists", listId);
     }
 }
